+++ src/main/java/io/zipcoder/casino/Players/DicePlayer.java
@@ -2,7 +2,4 @@ package io.zipcoder.casino.Players;
 
 public abstract class DicePlayer extends Player {
 
-    public DicePlayer(Profile profile) {
-        super(profile);
-    }
 }
