+++ src/main/java/io/zipcoder/casino/utilities/Greeter.java
@@ -1,7 +1,6 @@
 package io.zipcoder.casino.utilities;
 
 import io.zipcoder.casino.Cards.Games.*;
-import io.zipcoder.casino.utilities.Console;
 
 import java.util.NoSuchElementException;
